6. 6.24 : Bugs left for saving system : units garrison in buildings will need to be clicked twice to remove them when next time loaded. And unit click building to garrison failed.
7. 6.25 : Teest army group system with saving system.

+++ Assets/Scripts/Systems/General/BasicControl/Save/LoadSystemPlus.cs
@@ -1,50 +1,55 @@
-﻿using SparFlame.Components.General;
+﻿using System.IO;
+using SparFlame.Components.General;
+using SparFlame.Components.MainGameplay;
 using SparFlame.Components.SubGameplay;
 using Unity.Burst;
 using Unity.Collections;
 using Unity.Entities;
 using Unity.Entities.Serialization;
 using Unity.Physics;
-using UnityEngine;
 
 namespace SparFlame.Systems.General.BasicControl
 {
     [UpdateInGroup(typeof(InitializationSystemGroup))]
     public partial class LoadSystemPlus : SystemBase
     {
+        private EntityQuery _loadArmyGroupQuery;
+
         private NativeHashMap<int, Entity> _globalIdxToPrefabs;
         private NativeHashMap<long, Entity> _tmpIdxToInstances;
         private NativeHashMap<int, ExpStaticConfig> _expDatabase;
 
-        // private ComponentLookup<SeInGarrison> _inGarrisonLookup;
-        // private ComponentLookup<SeTmpId> _tmpIdLookup;
-        // private ComponentLookup<PhysicsMass> _physicsMassLookup;
-        // private ComponentLookup<SeInverseMass> _inverseMassLookup;
-        // private ComponentLookup<ExpData> _expDataLookup;
-        //
-        //
-        // private BufferLookup<SeGarrisonEntity> _garrisonEntitiesLookup;
-        // private BufferLookup<GarrisonTypeData> _garrisonTypeDataLookup;
+        private ComponentLookup<SeInGarrison> _inGarrisonLookup;
+        private ComponentLookup<SeTmpId> _tmpIdLookup;
+        private ComponentLookup<PhysicsMass> _physicsMassLookup;
+        private ComponentLookup<SeInverseMass> _inverseMassLookup;
+        private ComponentLookup<ExpData> _expDataLookup;
         
+        private BufferLookup<SeGarrisonEntity> _garrisonEntitiesLookup;
+        private BufferLookup<GarrisonTypeData> _garrisonTypeDataLookup;
+
         private ComponentLookup<VolumeObstacleSpawnRequest> _volumeObstacleSpawnRequestsLookup;
 
         private bool _initialized;
+
+
         protected override void OnCreate()
         {
-            RequireForUpdate<SaveConfig>();
-            RequireForUpdate<LoadData>();
+            RequireForUpdate<SaveLoadConfig>();
             RequireForUpdate<BuildingEntityPrefabData>();
             RequireForUpdate<UnitEntityPrefabData>();
             RequireForUpdate<PlayerSaveSlot>();
-            // _inGarrisonLookup = GetComponentLookup<SeInGarrison>(true);
-            // _inverseMassLookup = GetComponentLookup<SeInverseMass>(true);
-            // _tmpIdLookup = GetComponentLookup<SeTmpId>(true);
-            // _physicsMassLookup = GetComponentLookup<PhysicsMass>(true);
-            // _expDataLookup = GetComponentLookup<ExpData>(true);
-            //
-            // _garrisonEntitiesLookup = GetBufferLookup<SeGarrisonEntity>(true);
-            // _garrisonTypeDataLookup = GetBufferLookup<GarrisonTypeData>(true);
+            _inGarrisonLookup = GetComponentLookup<SeInGarrison>(true);
+            _inverseMassLookup = GetComponentLookup<SeInverseMass>(true);
+            _tmpIdLookup = GetComponentLookup<SeTmpId>(true);
+            _physicsMassLookup = GetComponentLookup<PhysicsMass>(true);
+            _expDataLookup = GetComponentLookup<ExpData>(true);
+            
+            _garrisonEntitiesLookup = GetBufferLookup<SeGarrisonEntity>(true);
+            _garrisonTypeDataLookup = GetBufferLookup<GarrisonTypeData>(true);
             _volumeObstacleSpawnRequestsLookup = GetComponentLookup<VolumeObstacleSpawnRequest>(true);
+
+            _loadArmyGroupQuery = SystemAPI.QueryBuilder().WithAll<InSubGameTag>().WithAll<ArmyGroupAttr>().Build();
         }
 
         protected override void OnDestroy()
@@ -62,52 +67,80 @@ namespace SparFlame.Systems.General.BasicControl
             if (!_initialized)
             {
                 Initialize();
-                GameController.Instance.OnEcsLoadCitySaving += LoadCity;
                 _initialized = true;
+                SaveLoadController.Instance.OnEcsLoadCitySubData += LoadCitySubData;
+                SaveLoadController.Instance.OnEcsLoadArmyGroupSubData += LoadArmyGroupSubData;
             }
-            
         }
 
         protected override void OnUpdate()
         {
         }
 
-        private void LoadCity(int cityId)
+        private void LoadArmyGroupSubData()
         {
             var playerSaveSlot = SystemAPI.GetSingleton<PlayerSaveSlot>().Value;
-            var ecb = new EntityCommandBuffer(Allocator.TempJob);
-            using (var deserializeWorld = new World("Deserialization World"))
+            var armyGroupAttrs = _loadArmyGroupQuery.ToComponentDataArray<ArmyGroupAttr>(Allocator.Temp);
+            foreach (var armyGroupAttr in armyGroupAttrs)
             {
-                var transaction = deserializeWorld.EntityManager.BeginExclusiveEntityTransaction();
-                using (var reader =
-                       new StreamBinaryReader(SaveUtilities.GetCitySavePath(cityId,playerSaveSlot)))
+                var armyGroupPath = SaveUtilities.GetArmyGroupSavePath(armyGroupAttr.SaveId, playerSaveSlot);
+                if(!File.Exists(armyGroupPath))continue;
+                using (var deserializeWorld = new World("Deserialization World"))
                 {
-                    SerializeUtility.DeserializeWorld(transaction, reader);
+                    var transaction = deserializeWorld.EntityManager.BeginExclusiveEntityTransaction();
+                    using (var reader =
+                           new StreamBinaryReader(armyGroupPath))
+                    {
+                        SerializeUtility.DeserializeWorld(transaction, reader);
+                    }
+
+                    deserializeWorld.EntityManager.EndExclusiveEntityTransaction();
+                    EntityManager.MoveEntitiesFrom(deserializeWorld.EntityManager);
+                    EntityManager.DestroyEntity(SystemAPI.GetSingletonEntity<SaveSystemPlus.SaveTmpTag>());
                 }
-                
-                deserializeWorld.EntityManager.EndExclusiveEntityTransaction();
-                EntityManager.MoveEntitiesFrom(deserializeWorld.EntityManager);
-                EntityManager.DestroyEntity(SystemAPI.GetSingletonEntity<SaveSystemPlus.SaveTmpTag>());
             }
-            // _expDataLookup.Update(this);
-            // _inverseMassLookup.Update(this);
-            // _physicsMassLookup.Update(this);
-            // _tmpIdLookup.Update(this);
-            // _inGarrisonLookup.Update(this);
-            //
-            // _garrisonEntitiesLookup.Update(this);
-            // _garrisonTypeDataLookup.Update(this);
-            _volumeObstacleSpawnRequestsLookup.Update(this);
-            var job =new BuildingLoadJob
+
+            armyGroupAttrs.Dispose();
+        }
+
+        private void LoadCitySubData(SubGameStatusData targetSubGameStatusData)
+        {
+            var city = targetSubGameStatusData.City;
+            var cityAttr = SystemAPI.GetComponent<CityAttr>(city);
+            var playerSaveSlot = SystemAPI.GetSingleton<PlayerSaveSlot>().Value;
+            var citySavePath = SaveUtilities.GetCitySavePath(cityAttr.ID, playerSaveSlot);
+
+            // Load city data
+            var fileExist = File.Exists(citySavePath);
+            if (fileExist)
             {
-                ECB = ecb.AsParallelWriter(),
-                Prefabs = _globalIdxToPrefabs,
-                RequestLookup = _volumeObstacleSpawnRequestsLookup,
-            }.ScheduleParallel(Dependency);
-            job.Complete();
-            ecb.Playback(EntityManager);
-            ecb.Dispose();
-            /*
+                using (var deserializeWorld = new World("Deserialization World"))
+                {
+                    var transaction = deserializeWorld.EntityManager.BeginExclusiveEntityTransaction();
+                    using (var reader =
+                           new StreamBinaryReader(citySavePath))
+                    {
+                        SerializeUtility.DeserializeWorld(transaction, reader);
+                    }
+
+                    deserializeWorld.EntityManager.EndExclusiveEntityTransaction();
+                    EntityManager.MoveEntitiesFrom(deserializeWorld.EntityManager);
+                    EntityManager.DestroyEntity(SystemAPI.GetSingletonEntity<SaveSystemPlus.SaveTmpTag>());
+                }
+            }
+           
+
+            _expDataLookup.Update(this);
+            _inverseMassLookup.Update(this);
+            _physicsMassLookup.Update(this);
+            _tmpIdLookup.Update(this);
+            _inGarrisonLookup.Update(this);
+            
+            _garrisonEntitiesLookup.Update(this);
+            _garrisonTypeDataLookup.Update(this);
+            var ecb = new EntityCommandBuffer(Allocator.TempJob);
+            
+            
             var loadJob = new LoadSubGameplayJob
             {
                 ECB = ecb.AsParallelWriter(),
@@ -127,28 +160,36 @@ namespace SparFlame.Systems.General.BasicControl
             {
                 _tmpIdxToInstances.Add(tmpId.ValueRO.value, entity);
             }
-
+            
             _garrisonEntitiesLookup.Update(this);
             _inGarrisonLookup.Update(this);
             var ecb2 = new EntityCommandBuffer(Allocator.TempJob);
-            var replaceJob = new ReplaceTmpIdJob
+            var replaceTmpRefJob = new ReplaceTmpIdJob
             {
                 ECB = ecb2.AsParallelWriter(),
                 SeGarrisonEntitiesLookup = _garrisonEntitiesLookup,
                 SeInGarrisonLookup = _inGarrisonLookup,
                 TmpIdxToInstances = _tmpIdxToInstances,
             }.ScheduleParallel(Dependency);
-            replaceJob.Complete();
+            replaceTmpRefJob.Complete();
             ecb2.Playback(EntityManager);
             ecb2.Dispose();
             _tmpIdxToInstances.Clear();
-            */
-
-            SystemAPI.SetSingleton(new LoadData
+            
+            
+            var ecb3 = new EntityCommandBuffer(Allocator.TempJob);
+            _volumeObstacleSpawnRequestsLookup.Update(this);
+            var buildingSpawnObstacleJob = new BuildingSpawnObstacleJob
             {
-                Type = SaveLoadType.None,
-                CityId = -1
-            });
+                ECB = ecb3.AsParallelWriter(),
+                Prefabs = _globalIdxToPrefabs,
+                RequestLookup = _volumeObstacleSpawnRequestsLookup,
+            }.ScheduleParallel(Dependency);
+            buildingSpawnObstacleJob.Complete();
+            ecb3.Playback(EntityManager);
+            ecb3.Dispose();
+
+         
         }
 
         private void Initialize()
@@ -176,18 +217,20 @@ namespace SparFlame.Systems.General.BasicControl
 
             _tmpIdxToInstances = new NativeHashMap<long, Entity>(100, allocator: Allocator.Persistent);
         }
-        
-        
+
+
         [BurstCompile]
         [WithNone(typeof(VolumeObstacleSpawnRequest))]
         [WithAll(typeof(BuildingAttr))]
         [WithNone(typeof(ConstructingData))]
-        public partial struct BuildingLoadJob : IJobEntity
+        public partial struct BuildingSpawnObstacleJob : IJobEntity
         {
             public EntityCommandBuffer.ParallelWriter ECB;
             [ReadOnly] public NativeHashMap<int, Entity> Prefabs;
             [ReadOnly] public ComponentLookup<VolumeObstacleSpawnRequest> RequestLookup;
-            private void Execute([ChunkIndexInQuery]int index,in SubGameplayGeneralAttr generalAttr, Entity selfEntity)
+
+            private void Execute([ChunkIndexInQuery] int index, in SubGameplayGeneralAttr generalAttr,
+                Entity selfEntity)
             {
                 var prefab = Prefabs[generalAttr.ID];
                 var request = RequestLookup[prefab];
